+++ postprocessors/unique_stock_ticker/unique_stock_ticker/clients/six/schema/common/lookup.py
@@ -1,0 +1,8 @@
+from enum import StrEnum
+
+
+class LookupStatus(StrEnum):
+    FOUND = "FOUND"
+    NOT_FOUND = "NOT_FOUND"
+    NOT_AUTHORIZED = "NOT_AUTHORIZED"
+    NOT_VALID_ID = "NOT_VALID_ID"