+++ postprocessors/unique_stock_ticker/unique_stock_ticker/clients/six/schema/free_text_search/entities/request.py
@@ -1,0 +1,7 @@
+from unique_stock_ticker.clients.six.schema import BaseRequestParams, EntityType
+
+
+class FreeTextSearchEntitiesRequestParams(BaseRequestParams):
+    text: str
+    size: int | None = 10
+    entity_type: EntityType | None = None