+++ unique_toolkit/unique_toolkit/_common/validators.py
@@ -1,12 +1,16 @@
-from typing import Annotated
+import logging
+from typing import Annotated, Any
 
-from pydantic import BeforeValidator, PlainSerializer
+from pydantic import BeforeValidator, Field, PlainSerializer, ValidationInfo
+from pydantic.fields import FieldInfo
 
 from unique_toolkit.language_model import LanguageModelName
 from unique_toolkit.language_model.infos import (
     LanguageModelInfo,
     LanguageModelProvider,
 )
+
+logger = logging.getLogger(__name__)
 
 # TODO @klcd: Inform on deprecation of str as input
 LMI = Annotated[
@@ -21,6 +25,14 @@ LMI = Annotated[
         return_type=str | LanguageModelInfo,
     ),
 ]
+
+
+def get_LMI_default_field(llm_name: LanguageModelName, **kwargs) -> Any:
+    return Field(
+        default=LanguageModelInfo.from_name(llm_name),
+        json_schema_extra={"default": llm_name},
+        **kwargs,
+    )
 
 
 def serialize_lmi(model: LanguageModelInfo) -> str | LanguageModelInfo:
@@ -55,3 +67,32 @@ def validate_and_init_language_model_info(
         )
 
     return v
+
+
+def ClipInt(*, min_value: int, max_value: int) -> tuple[BeforeValidator, FieldInfo]:
+    def _validator(value: Any, info: ValidationInfo) -> Any:
+        if not isinstance(value, int):
+            value = int(value)
+
+        field_name = info.field_name
+        if value < min_value:
+            logger.warning(
+                "Field %s is below the allowed minimum of %s. It will be set to %s.",
+                field_name,
+                min_value,
+                min_value,
+            )
+            return min_value
+
+        if value > max_value:
+            logger.warning(
+                "Field %s is above the allowed maximum of %s. It will be set to %s.",
+                field_name,
+                max_value,
+                max_value,
+            )
+            return max_value
+
+        return value
+
+    return (BeforeValidator(_validator), Field(ge=min_value, le=max_value))