+++ unique_toolkit/unique_toolkit/tools/tool_manager.py
@@ -1,9 +1,14 @@
 import asyncio
 from logging import Logger, getLogger
+from typing import Any
+
 from pydantic import BaseModel, Field
+
 from unique_toolkit.app.schemas import ChatEvent
+from unique_toolkit.evals.schemas import EvaluationMetricName
 from unique_toolkit.language_model.schemas import (
     LanguageModelFunction,
+    LanguageModelTool,
     LanguageModelToolDescription,
 )
 from unique_toolkit.tools.config import ToolBuildConfig
@@ -12,13 +17,6 @@ from unique_toolkit.tools.schemas import ToolCallResponse, ToolPrompts
 from unique_toolkit.tools.tool import Tool
 from unique_toolkit.tools.tool_progress_reporter import ToolProgressReporter
 from unique_toolkit.tools.utils.execution.execution import Result, SafeTaskExecutor
-
-
-class ForcedToolOption:
-    type: str = "function"
-
-    def __init__(self, name: str):
-        self.name = name
 
 
 class ToolManagerConfig(BaseModel):
@@ -32,10 +30,6 @@ class ToolManagerConfig(BaseModel):
         ge=1,
         description="Maximum number of tool calls that can be executed in one iteration.",
     )
-
-    def __init__(self, tools: list[ToolBuildConfig], max_tool_calls: int = 10):
-        self.tools = tools
-        self.max_tool_calls = max_tool_calls
 
 
 class ToolManager:
@@ -67,16 +61,18 @@ class ToolManager:
         tool_progress_reporter: ToolProgressReporter,
     ):
         self._logger = logger
-        self._config  = config
+        self._config = config
         self._tool_progress_reporter = tool_progress_reporter
         self._tools = []
         self._tool_choices = event.payload.tool_choices
         self._disabled_tools = event.payload.disabled_tools
+        # this needs to be a set of strings to avoid duplicates
+        self._tool_evaluation_check_list: set[EvaluationMetricName] = set()
         self._init__tools(event)
 
-    def _init__tools(self,  event: ChatEvent) -> None:
+    def _init__tools(self, event: ChatEvent) -> None:
         tool_choices = self._tool_choices
-        tool_configs = self._config .tools
+        tool_configs = self._config.tools
         self._logger.info("Initializing tool definitions...")
         self._logger.info(f"Tool choices: {tool_choices}")
         self._logger.info(f"Tool configs: {tool_configs}")
@@ -105,6 +101,9 @@ class ToolManager:
 
             self._tools.append(t)
 
+    def get_evaluation_check_list(self) -> list[EvaluationMetricName]:
+        return list(self._tool_evaluation_check_list)
+
     def log_loaded_tools(self):
         self._logger.info(f"Loaded tools: {[tool.name for tool in self._tools]}")
 
@@ -117,10 +116,23 @@ class ToolManager:
                 return tool
         return None
 
-    def get_forced_tools(self) -> list[ForcedToolOption]:
-        return [ForcedToolOption(t.name) for t in self._tools if t.name in self._tool_choices]
+    def get_forced_tools(self) -> list[dict[str, Any]]:
+        return [
+            self._convert_to_forced_tool(t.name)
+            for t in self._tools
+            if t.name in self._tool_choices
+        ]
 
-    def get_tool_definitions(self) -> list[LanguageModelToolDescription]:
+    def add_forced_tool(self, name):
+        tool = self.get_tool_by_name(name)
+        if not tool:
+            raise ValueError(f"Tool {name} not found")
+        self._tools.append(tool)
+        self._tool_choices.append(tool.name)
+
+    def get_tool_definitions(
+        self,
+    ) -> list[LanguageModelTool | LanguageModelToolDescription]:
         return [tool.tool_description() for tool in self._tools]
 
     def get_tool_prompts(self) -> list[ToolPrompts]:
@@ -137,19 +149,19 @@ class ToolManager:
         )
         num_tool_calls = len(tool_calls)
 
-        if num_tool_calls > self._config .max_tool_calls:
+        if num_tool_calls > self._config.max_tool_calls:
             self._logger.warning(
                 (
                     "Number of tool calls %s exceeds the allowed maximum of %s."
                     "The tool calls will be reduced to the first %s."
                 ),
                 num_tool_calls,
-                self._config .max_tool_calls,
-                self._config .max_tool_calls,
+                self._config.max_tool_calls,
+                self._config.max_tool_calls,
             )
-            tool_calls = tool_calls[: self._config .max_tool_calls]
+            tool_calls = tool_calls[: self._config.max_tool_calls]
 
-        tool_call_responses = await self._execute_parallelized(tool_calls=tool_calls)
+        tool_call_responses = await self._execute_parallelized(tool_calls)
         return tool_call_responses
 
     async def _execute_parallelized(
@@ -187,13 +199,18 @@ class ToolManager:
     ) -> ToolCallResponse:
         self._logger.info(f"Processing tool call: {tool_call.name}")
 
-        tool_instance = self.get_tool_by_name(tool_call.name)
+        tool_instance = self.get_tool_by_name(
+            tool_call.name
+        )  # we need to copy this as it will have problematic interference on multi calls.
 
         if tool_instance:
             # Execute the tool
             tool_response: ToolCallResponse = await tool_instance.run(
                 tool_call=tool_call
             )
+            evaluation_checks = tool_instance.evaluation_check_list()
+            self._tool_evaluation_check_list.update(evaluation_checks)
+
             return tool_response
 
         return ToolCallResponse(
@@ -240,3 +257,11 @@ class ToolManager:
                 f"Filtered out {len(tool_calls) - len(unique_tool_calls)} duplicate tool calls."
             )
         return unique_tool_calls
+
+    from typing import Any
+
+    def _convert_to_forced_tool(self, tool_name: str) -> dict[str, Any]:
+        return {
+            "type": "function",
+            "function": {"name": tool_name},
+        }