+++ unique_toolkit/unique_toolkit/tools/utils/source_handling/tests/test_source_formatting.py
@@ -1,4 +1,5 @@
 import pytest
+
 from unique_toolkit.content.schemas import ContentChunk
 from unique_toolkit.tools.utils.source_handling.schema import SourceFormatConfig
 from unique_toolkit.tools.utils.source_handling.source_formatting import (